*** RELEASE 0.1 ***
Initial public release.  CLI scripts can be run through this system (via cron) 
and their results are logged to a database.  This system is eventually intended 
as a way of supplementing a cron daemon, running as its own daemon, so jobs can 
be added, modified, and deleted in an online format without having direct access 
to the server itself.

The present release deals simply with getting the script output into a database 
and separating error from output.